Fix MethodNotFoundException when applying manual TMDB search results outside List pages
The manual_tmdb_search action (VodResource/SeriesResource table actions)
renders the tmdb-search-results component, whose result buttons call
applyTmdbSelection on the hosting Livewire component. The handler and its
two helpers only existed on ListVod/ListSeries, so applying a result from
any other hosting component (the VOD/Series view pages, or the relation
managers on Custom Playlists and Media Server Integrations that reuse
getTableActions()) threw:

  Livewire\Exceptions\MethodNotFoundException: Unable to call component
  method. Public method [applyTmdbSelection] not found on component

The two List-page copies were byte-identical, so extract them into a new
App\Traits\AppliesTmdbSelection trait and use it in every component that
can mount the action: ListVod, ListSeries, ViewVod, ViewSeries,
VodRelationManager, MoviesRelationManager and both SeriesRelationManagers.
