+++ app/Filament/Resources/MediaServerIntegrations/RelationManagers/MoviesRelationManager.php
@@ -3,6 +3,7 @@
 namespace App\Filament\Resources\MediaServerIntegrations\RelationManagers;
 
 use App\Filament\Resources\Vods\VodResource;
+use App\Traits\AppliesTmdbSelection;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Schemas\Components\Tabs\Tab;
 use Filament\Schemas\Schema;
@@ -13,6 +14,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class MoviesRelationManager extends RelationManager
 {
+    use AppliesTmdbSelection;
+
     protected static string $relationship = 'channels';
 
     protected static ?string $label = 'Movies';
